+++ silkopter/libs/common/stream/ICurrent.h
@@ -12,7 +12,7 @@ class ICurrent : public IScalar_Stream<Type::CURRENT>
 public:
     typedef std::true_type can_be_filtered_t;
 
-    typedef double             Value; //amperes
+    typedef float             Value; //amperes
     typedef stream::Sample<Value>     Sample;
     virtual auto get_samples() const -> std::vector<Sample> const& = 0;
 };
