+++ silkopter/libs/common/stream/IPressure.h
@@ -12,7 +12,7 @@ class IPressure : public IScalar_Stream<Type::PRESSURE>
 public:
     typedef std::true_type can_be_filtered_t;
 
-    typedef double                   Value; //??
+    typedef float                   Value; //??
     typedef stream::Sample<Value>     Sample;
     virtual auto get_samples() const -> std::vector<Sample> const& = 0;
 };
