+++ silkopter/libs/common/stream/ITemperature.h
@@ -12,7 +12,7 @@ class ITemperature : public IScalar_Stream<Type::TEMPERATURE>
 public:
     typedef std::true_type can_be_filtered_t;
 
-    typedef double                   Value; //degrees celsius
+    typedef float                   Value; //degrees celsius
     typedef stream::Sample<Value>     Sample;
     virtual auto get_samples() const -> std::vector<Sample> const& = 0;
 };
